# [pkg/stanza] improve performance