--- a/pkg/stanza/operator/helper/encoding.go
+++ b/pkg/stanza/operator/helper/encoding.go
@@ -17,12 +17,12 @@ package helper // import "github.com/open-telemetry/opentelemetry-collector-cont
 import (
 	"errors"
 	"fmt"
+	"golang.org/x/text/transform"
 	"strings"
 
 	"golang.org/x/text/encoding"
 	"golang.org/x/text/encoding/ianaindex"
 	"golang.org/x/text/encoding/unicode"
-	"golang.org/x/text/transform"
 )
 
 // NewBasicConfig creates a new Encoding config
@@ -45,27 +45,28 @@ func (c EncodingConfig) Build() (Encoding, error) {
 	}
 
 	return Encoding{
-		Encoding: enc,
+		Encoding:     enc,
+		decodeBuffer: make([]byte, 1<<12),
+		decoder:      enc.NewDecoder(),
 	}, nil
 }
 
 type Encoding struct {
-	Encoding encoding.Encoding
+	Encoding     encoding.Encoding
+	decoder      *encoding.Decoder
+	decodeBuffer []byte
 }
 
-// decode converts the bytes in msgBuf to utf-8 from the configured encoding
+// Decode converts the bytes in msgBuf to utf-8 from the configured encoding
 func (e *Encoding) Decode(msgBuf []byte) ([]byte, error) {
-	decodeBuffer := make([]byte, 1<<12)
-	decoder := e.Encoding.NewDecoder()
-
 	for {
-		decoder.Reset()
-		nDst, _, err := decoder.Transform(decodeBuffer, msgBuf, true)
+		e.decoder.Reset()
+		nDst, _, err := e.decoder.Transform(e.decodeBuffer, msgBuf, true)
 		if err == nil {
-			return decodeBuffer[:nDst], nil
+			return e.decodeBuffer[:nDst], nil
 		}
 		if errors.Is(err, transform.ErrShortDst) {
-			decodeBuffer = make([]byte, len(decodeBuffer)*2)
+			e.decodeBuffer = make([]byte, len(e.decodeBuffer)*2)
 			continue
 		}
 		return nil, fmt.Errorf("transform encoding: %w", err)